splitting controllers and service classes

+++ src/main/java/com/example/repository/CityRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import com.example.entity.City;
 
-public interface CityRepo extends JpaRepository<City, Integer> {
+public interface CityRepo extends JpaRepository<City, Long> {
 //	@Query("select c from City c where c.country.name = ?1")
 	List<City> findByCountry(String country);
 }
