+++ src/main/java/com/example/repository/NeighbourhoodRepo.java
@@ -3,6 +3,6 @@ package com.example.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.entity.Neighbourhood;
 
-public interface NeighbourhoodRepo extends JpaRepository<Neighbourhood, Integer> {
+public interface NeighbourhoodRepo extends JpaRepository<Neighbourhood, Long> {
 
 }
